Changed kill tallies to use persistent methods, changed worst-performing clan algorithm
When server tallies kills for each clan, the server will use persistent
Civilian/Neutral/Rival methods that are saved in the database so that
servers that stay up for an extended period of time won't be affected by
a server crash. Also changed the algorithm that determines the
worst-performing clan by comparing deaths in addition to kills.